Move extension into base library
Required for debugging specific handlers
